+++ packages/create-app/src/enums/template.ts
@@ -1,0 +1,16 @@
+/*
+ * Copyright (c) 2024 Kiyozz.
+ *
+ * All rights reserved.
+ */
+
+export enum Template {
+  ReactTypeScript = 'react-ts',
+  ReactJavaScript = 'react',
+  ReactViteJavaScript = 'react-vite',
+  ReactViteTypeScript = 'react-vite-ts',
+  MainOnlyTypeScript = 'main-ts',
+  MainOnlyJavaScript = 'main',
+  SvelteJavaScript = 'svelte',
+  SvelteTypeScript = 'svelte-ts',
+}