+++ packages/create-app/templates/main/src/main/main.js
@@ -1,9 +1,12 @@
 import { app, BrowserWindow } from 'electron'
+import { isDev } from 'electron-util/main'
 import { is } from 'electron-util'
-import * as path from 'path'
+import * as path from 'node:path'
 
 /** @type {BrowserWindow | null} */
 let win = null
+
+const dirname = path.dirname(new URL(import.meta.url).pathname)
 
 async function createWindow() {
   win = new BrowserWindow({
@@ -12,12 +15,10 @@ async function createWindow() {
     minHeight: 600,
     minWidth: 650,
     webPreferences: {
-      preload: path.join(__dirname, 'preload.js'),
+      preload: path.join(dirname, 'preload.js'),
     },
     show: false,
   })
-
-  const isDev = is.development
 
   win.loadURL('https://duckduckgo.com') // load any url you want to use with electron
 
@@ -39,7 +40,7 @@ async function createWindow() {
   })
 }
 
-app.on('ready', createWindow)
+app.whenReady().then(createWindow)
 
 app.on('window-all-closed', () => {
   if (!is.macos) {