+++ packages/create-app/templates/react-ts/esbuild.main.config.ts
@@ -1,5 +1,5 @@
 import { BuildOptions } from 'esbuild'
-import * as path from 'path'
+import * as path from 'node:path'
 
 const config: BuildOptions = {
   platform: 'node',
@@ -8,7 +8,8 @@ const config: BuildOptions = {
     path.resolve('src/main/preload.ts'),
   ],
   bundle: true,
-  target: 'node18.15.0', // electron version target
+  format: 'esm',
+  target: 'node20.11.1', // electron version target
 }
 
 export default config