+++ packages/create-app/templates/react-vite-ts/src/renderer/vite.config.ts
@@ -1,11 +1,11 @@
 import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
+import react from '@vitejs/plugin-react-swc'
 
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react()],
   build: {
-    target: 'chrome114', // electron version target
+    target: 'chrome124', // electron version target
   },
   server: {
     port: 9080,