+++ packages/create-app/templates/svelte-ts/esbuild.main.config.ts
@@ -1,14 +1,22 @@
 import { BuildOptions } from 'esbuild'
-import * as path from 'path'
+import * as path from 'node:path'
 
 const config: BuildOptions = {
   platform: 'node',
-  entryPoints: [
-    path.resolve('src/main/main.ts'),
-    path.resolve('src/main/preload.ts'),
-  ],
+  // entryPoints: [
+  //   path.resolve('src/main/main.ts'),
+  //   path.resolve('src/main/preload.ts'),
+  // ],
+  entryPoints: {
+    'main': path.resolve('src/main/main.ts'),
+    'preload': path.resolve('src/main/preload.mts')
+  },
+  outExtension: {
+    '.js': '.mjs'
+  },
   bundle: true,
-  target: 'node18.15.0', // electron version target
+  format: 'esm',
+  target: 'node20.11.1', // electron version target
 }
 
 export default config