+++ packages/electron-esbuild/src/env.ts
@@ -1,0 +1,7 @@
+/*
+ * Copyright (c) 2024 Kiyozz.
+ *
+ * All rights reserved.
+ */
+
+export type Env = 'production' | 'development'