+++ packages/create-app/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 9.0.0
+
+### Breaking Changes
+
+- change to pure ESM
+
+### Changes
+
+- Updated all templates for latest electron versions.
+
 ## 8.0.0
 
 ### Breaking Changes